+++ schematics/ng-update/update-9/migrations/import-entry-point-trim.ts
@@ -1,4 +1,4 @@
-import ts from 'typescript';
+import ts, { factory } from 'typescript';
 import { ContentChange, ReplaceContentChange } from '../../../types';
 import { MigrationBase } from './base';
 
@@ -9,16 +9,20 @@ export class ImportEntryPointTrimMigration extends MigrationBase {
         const contentChangeList: ContentChange[] = [];
 
         for (const importDeclaration of importDeclarationList) {
+            let tempImportDeclaration: ts.ImportDeclaration;
             const packagePathList = this.getImportDeclarationPackageName(importDeclaration)
                 .split('/')
                 .filter(item => item);
             if (packagePathList.length > 2) {
-                importDeclaration.moduleSpecifier = this.createStringLiteral(packagePathList.slice(0, 2).join('/'));
+                tempImportDeclaration = this.updateImportDeclaration(
+                    importDeclaration,
+                    this.createStringLiteral(packagePathList.slice(0, 2).join('/'))
+                );
                 contentChangeList.push(
                     new ReplaceContentChange(
                         importDeclaration.getStart(),
                         importDeclaration.getWidth(),
-                        this.printNodeContent(importDeclaration)
+                        this.printNodeContent(tempImportDeclaration)
                     )
                 );
             }