+++ site/src/app/app.module.ts
@@ -1,49 +1,15 @@
-import {
-    DocgeniTemplateModule,
-    CONFIG_TOKEN,
-    routes,
-    RootComponent,
-    initializeDocgeniSite,
-    GlobalContext,
-    NavigationService,
-    PageTitleService,
-    CopierService
-} from '@docgeni/template';
-import { NgModule, APP_INITIALIZER } from '@angular/core';
-import { config } from './content/config';
+import { NgModule } from '@angular/core';
 import { RouterModule } from '@angular/router';
-import { LIB_EXAMPLE_LOADER_PROVIDER } from './content/example-loader';
-import { EXAMPLE_MODULES } from './content/example-modules';
-import './content/navigations.json';
 import { AppComponent } from './app.component';
 import { NgxTethysModule } from 'ngx-tethys';
 import { BrowserModule } from '@angular/platform-browser';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-
+import { EXAMPLE_MODULES, DOCGENI_SITE_PROVIDERS } from './content/index';
+import { DocgeniTemplateModule } from '@docgeni/template';
 @NgModule({
     declarations: [AppComponent],
-    imports: [
-        BrowserModule,
-        BrowserAnimationsModule,
-        DocgeniTemplateModule,
-        RouterModule.forRoot(routes),
-        NgxTethysModule,
-        ...EXAMPLE_MODULES
-    ],
-    providers: [
-        BrowserModule,
-        BrowserAnimationsModule,
-        GlobalContext,
-        NavigationService,
-        PageTitleService,
-        CopierService,
-        { provide: APP_INITIALIZER, useFactory: initializeDocgeniSite, deps: [GlobalContext], multi: true },
-        LIB_EXAMPLE_LOADER_PROVIDER,
-        {
-            provide: CONFIG_TOKEN,
-            useValue: config
-        }
-    ],
+    imports: [BrowserModule, BrowserAnimationsModule, DocgeniTemplateModule, RouterModule.forRoot([]), NgxTethysModule, ...EXAMPLE_MODULES],
+    providers: [...DOCGENI_SITE_PROVIDERS],
     bootstrap: [AppComponent]
 })
 export class AppModule {