+++ src/uploader/examples/module.ts
@@ -1,6 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { ThyUploaderModule, ThyProgressModule, ThyButtonModule } from 'ngx-tethys';
+import { FormsModule } from '@angular/forms';
+import { ThyUploaderModule } from 'ngx-tethys/uploader';
+import { ThyProgressModule } from 'ngx-tethys/progress';
+import { ThyButtonModule } from 'ngx-tethys/button';
+import { ThyFormModule } from 'ngx-tethys/form';
+import { ThyInputModule } from 'ngx-tethys/input';
+import { ThyCheckboxModule } from 'ngx-tethys/checkbox';
 import { ThyUploaderBasicExampleComponent } from './basic/basic.component';
 import { ThyUploaderMultipleExampleComponent } from './multiple/multiple.component';
 import { ThyUploaderAcceptFolderExampleComponent } from './accept-folder/accept-folder.component';
@@ -16,7 +22,16 @@ const COMPONENTS = [
 @NgModule({
     declarations: [...COMPONENTS],
     entryComponents: [...COMPONENTS],
-    imports: [ThyUploaderModule, ThyProgressModule, CommonModule, ThyButtonModule],
+    imports: [
+        ThyUploaderModule,
+        ThyProgressModule,
+        ThyButtonModule,
+        ThyFormModule,
+        ThyInputModule,
+        ThyCheckboxModule,
+        CommonModule,
+        FormsModule
+    ],
     exports: [...COMPONENTS]
 })
 export class ThyUploaderExamplesModule {}